Implemented DailyPredictor and updated SMAPredictor.
Also added DiscreteDistribution to encapsulate a discrete distribution
where each element has a name.

Added BrokerInfoAccess so that predictors had limited access to the
broker: they should be able to get time and price info, but they shouldn't
be able to, for example, open a new account.

NOTE: RetireTool is currently broken!